requirements.txt
click
rich
pydantic
openai
pytest
python-dateutil
